pyproject.toml
[tool.poetry]
name = "dutchdesignbot"
version = "0.1.0"
description = ""
authors = ["bramiozo <bramiozo@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
PyYAML = "^6.0.2"
openai = "^1.50.1"
elevenlabs = "^1.9.0"
pydantic = "^2.9.2"
sentence-transformers = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
